• add a bundle description to the info.plist (using the key ‘description’)
There is no GUI editor for this and currently this is not presented anywhere.

I have used Markdown as the description format, the rationale being that we likely do want to embed links in the description, but we want to stay as close to plain text as possible, as there is no guarantee that the description goes through any preprocessing before it is being displayed.

git-svn-id: http://svn.textmate.org/trunk/Bundles/S5.tmbundle@4888 dfb7d73b-c2ec-0310-8fea-fb051d288c6d